+++ .gitignore
@@ -30,6 +30,7 @@ logs/
 .env.dev
 .pytest_cache/
 .ruff_cache/
+*.egg-info/
 
 # Misc
 .terraform/
