+++ src/devonboarder/__init__.py
@@ -1,0 +1,5 @@
+"""DevOnboarder package."""
+
+from .app import greet
+
+__all__ = ["greet"]
